Move function `canonicalize`, `apply` and `test` into `number.hpp`
