+++ README.md
@@ -119,16 +119,16 @@ sudo rm -rf /usr/local/share/meevax
 
 | Target Name        | Description
 |--------------------|---
-| `all` (default)    | Build shared-library `libmeevax.0.4.635.so` and executable `meevax`
+| `all` (default)    | Build shared-library `libmeevax.0.4.636.so` and executable `meevax`
 | `test`             | Test executable `meevax`
-| `package`          | Generate debian package `meevax_0.4.635_amd64.deb`
+| `package`          | Generate debian package `meevax_0.4.636_amd64.deb`
 | `install`          | Copy files into `/usr/local`
 | `install.deb`      | `all` + `package` + `sudo apt install <meevax>.deb`
 
 ## Usage
 
 ```
-Meevax Lisp 0.4.635
+Meevax Lisp 0.4.636
 
 Usage:
   meevax [option...] [file...]
