+++ include/meevax/kernel/number.hpp
@@ -329,6 +329,111 @@ inline namespace kernel
 
 inline namespace number
 {
+  template <typename T>
+  auto canonicalize(T&& x) -> decltype(auto)
+  {
+    if constexpr (std::is_same_v<std::decay_t<T>, object>)
+    {
+      return std::forward<decltype(x)>(x);
+    }
+    else if constexpr (std::is_same_v<std::decay_t<T>, complex>)
+    {
+      return x.canonicalize();
+    }
+    else if constexpr (std::is_same_v<std::decay_t<T>, ratio>)
+    {
+      return x.denominator() == 1 ? make(x.numerator()) : make(std::forward<decltype(x)>(x));
+    }
+    else
+    {
+      return make(std::forward<decltype(x)>(x));
+    }
+  }
+
+  template <auto I = 0, typename F>
+  auto apply([[maybe_unused]] F f, object const& x) -> object
+  {
+    using Ts = std::tuple<exact_integer, ratio, float, double, complex>;
+
+    if constexpr (I < std::tuple_size_v<Ts>)
+    {
+      using T = std::tuple_element_t<I, Ts>;
+
+      if (x.is<T>())
+      {
+        return canonicalize(f(x.as<T>()));
+      }
+      else
+      {
+        return apply<I + 1>(f, x);
+      }
+    }
+    else
+    {
+      throw std::domain_error("not an number");
+    }
+  }
+
+  template <auto I = 0, typename F>
+  auto apply([[maybe_unused]] F f, object const& x, object const& y) -> object
+  {
+    using Ts = combination<exact_integer, ratio, float, double, complex>;
+
+    if constexpr (I < std::tuple_size_v<Ts>)
+    {
+      using T = std::tuple_element_t<0, std::tuple_element_t<I, Ts>>;
+      using U = std::tuple_element_t<1, std::tuple_element_t<I, Ts>>;
+
+      if (x.is<T>() and y.is<U>())
+      {
+        return canonicalize(f(x.as<T>(), y.as<U>()));
+      }
+      else
+      {
+        return apply<I + 1>(f, x, y);
+      }
+    }
+    else
+    {
+      throw std::domain_error("not an number");
+    }
+  }
+
+  template <auto I = 0, typename F>
+  auto test([[maybe_unused]] F f, object const& x) -> bool
+  {
+    using Ts = std::tuple<exact_integer, ratio, float, double, complex>;
+
+    if constexpr (I < std::tuple_size_v<Ts>)
+    {
+      using T = std::tuple_element_t<I, Ts>;
+
+      return x.is<T>() ? f(x.as<T>()) : test<I + 1>(f, x);
+    }
+    else
+    {
+      return false;
+    }
+  }
+
+  template <auto I = 0, typename F>
+  auto test([[maybe_unused]] F f, object const& x, object const& y) -> bool
+  {
+    using Ts = combination<exact_integer, ratio, float, double, complex>;
+
+    if constexpr (I < std::tuple_size_v<Ts>)
+    {
+      using T = std::tuple_element_t<0, std::tuple_element_t<I, Ts>>;
+      using U = std::tuple_element_t<1, std::tuple_element_t<I, Ts>>;
+
+      return x.is<T>() and y.is<U>() ? f(x.as<T>(), y.as<U>()) : test<I + 1>(f, x, y);
+    }
+    else
+    {
+      return false;
+    }
+  }
+
   auto equals(object const&, object const&) -> bool;
 
   auto not_equals(object const&, object const&) -> bool;
